+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Change Log
 ==========
 
+ * renamed 'concatenateFiles' to 'concatenateFileArrays'
  * renamed 'concatenateBlobs' to 'concatenateBlobArrays'
  * extended 'javac' function by adding support to "1.8"/"8" values to 'source' and 'target' parameters
  * removed support for Windows OS